fix: set Base as default label
If an object is deleted, we will stop resolution of the object before the ObjectType is determined. The ObjectType will in this case be the default (first) Label value, User.
This resulted in the 'Deleted Objects' container (which is marked as deleted) being resolved to a deleted User.

I have changed the the default Label to Base.